+++ client/src/components/pages/projects/Wizard.js
@@ -414,7 +414,7 @@ class Wizard extends React.Component {
                       "us-east": ["us-east", "wdc06", "wdc07"],
                       "us-south": ["us-south", "dal10", "dal12"],
                       "jp-tok": ["tok04"],
-                      "br-sao": [], //["sao01", "sao04"], removed sao from power until public endpoint available
+                      "br-sao": ["sao01", "sao04"],
                       "ca-tor": ["tor01"],
                     }[this.state.region]
                   }
@@ -428,7 +428,7 @@ class Wizard extends React.Component {
                         "eu-gb",
                         "eu-es",
                         "us-east",
-                        //"br-sao", removed sao from power until public endpoint available
+                        "br-sao",
                         "jp-tok",
                         "ca-tor",
                       ],
@@ -444,7 +444,7 @@ class Wizard extends React.Component {
                         "eu-gb",
                         "eu-es",
                         "us-east",
-                        //"br-sao", removed sao from power until public endpoint available
+                        "br-sao",
                         "jp-tok",
                         "ca-tor",
                       ],